refactor the lp model (unique lp formulation but then possible to use mathopt/gurobi backend as done in the lib), minizinc flow solver
