+++ examples/multibatching/solvers_config.py
@@ -17,7 +17,7 @@ from discrete_optimization.multibatching.solvers.cpsat import (
     CpsatMultibatchingSolver,
     ModelingMultiBatch,
 )
-from discrete_optimization.multibatching.solvers.milp_flow import (
+from discrete_optimization.multibatching.solvers.lp import (
     GurobiMultibatchingSolver,
     GurobiMultibatchingSolverUnitFlow,
 )
