+++ src/discrete_optimization/multibatching/solvers/two_steps.py
@@ -20,7 +20,7 @@ from discrete_optimization.multibatching.solvers.cpsat import (
     CpsatMultibatchingSolver,
     ModelingMultiBatch,
 )
-from discrete_optimization.multibatching.solvers.milp_flow import (
+from discrete_optimization.multibatching.solvers.lp import (
     GurobiMultibatchingSolver,
 )
 from discrete_optimization.multibatching.solvers.netx import NetxMultibatchingSolver
